price-intelligence-system/requirements.txt
# Core Dependencies
python-dotenv==1.0.0

# Web Scraping
scrapy==2.11.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
fake-useragent==1.4.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Data Transformation (dbt)
dbt-core==1.7.4
dbt-postgres==1.7.4

# API
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
marshmallow==3.20.1

# Dashboard & Visualization
dash==2.14.2
plotly==5.18.0
dash-bootstrap-components==1.5.0

# Orchestration (for later)
# apache-airflow==2.7.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1

# Development & Testing
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
ipython==8.18.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk==1.39.1
